issue-29 Move classed to packages where they are used
