+++ src/main/kotlin/pullrequestfactory/io/clikt/OpenCommand.kt
@@ -4,7 +4,7 @@ import com.github.ajalt.clikt.core.CliktCommand
 import com.github.ajalt.clikt.core.context
 import com.github.ajalt.clikt.sources.PropertiesValueSource
 import pullrequestfactory.domain.Candidate
-import pullrequestfactory.io.GithubAPIClient
+import pullrequestfactory.io.programs.impl.GithubAPIClient
 import pullrequestfactory.io.programs.impl.OpenPullRequestsProgram
 import pullrequestfactory.io.repositories.KhttpClient
 import pullrequestfactory.io.repositories.KhttpClientStats
