+++ src/main/kotlin/pullrequestfactory/io/programs/impl/ClosePullRequestsPrograms.kt
@@ -2,7 +2,6 @@ package pullrequestfactory.io.programs.impl
 
 import pullrequestfactory.domain.Candidate
 import pullrequestfactory.domain.uis.UI
-import pullrequestfactory.io.GithubAPIClient
 import pullrequestfactory.io.programs.Program
 import pullrequestfactory.io.repositories.KhttpClientStats
