+++ src/main/kotlin/pullrequestfactory/io/programs/impl/DebugRateLimitCheckedProgram.kt
@@ -1,8 +1,6 @@
 package pullrequestfactory.io.programs.impl
 
 import pullrequestfactory.domain.uis.UI
-import pullrequestfactory.io.GithubAPIClient
-import pullrequestfactory.io.RateLimit
 import pullrequestfactory.io.programs.Program
 import pullrequestfactory.io.repositories.KhttpClientStats
 import java.lang.System.lineSeparator
