+++ src/main/kotlin/pullrequestfactory/io/programs/impl/Rate.kt
@@ -1,0 +1,5 @@
+package pullrequestfactory.io.programs.impl
+
+import java.time.Instant
+
+data class Rate(val limit: Int, val remaining: Int, val reset: Instant, val used: Int)
