+++ src/main/kotlin/pullrequestfactory/io/programs/impl/RateLimitCheckedProgram.kt
@@ -1,8 +1,6 @@
 package pullrequestfactory.io.programs.impl
 
 import pullrequestfactory.domain.uis.UI
-import pullrequestfactory.io.GithubAPIClient
-import pullrequestfactory.io.RateLimit
 import pullrequestfactory.io.programs.Program
 
 class RateLimitCheckedProgram(
