+++ src/main/kotlin/pullrequestfactory/io/programs/impl/RateLimitCheckedPrograms.kt
@@ -1,7 +1,6 @@
 package pullrequestfactory.io.programs.impl
 
 import pullrequestfactory.domain.uis.UI
-import pullrequestfactory.io.GithubAPIClient
 import pullrequestfactory.io.programs.Program
 import pullrequestfactory.io.repositories.KhttpClientStats
